+++ 3Week_Mission/src/main/java/com/project/Week_Mission/app/order/exception/OrderNotEnoughRestCashException.java
@@ -1,4 +1,4 @@
-package com.project.Week_Mission.app.order.controller;
+package com.project.Week_Mission.app.order.exception;
 
 public class OrderNotEnoughRestCashException extends RuntimeException {
 }
