+++ 3Week_Mission/src/main/java/com/project/Week_Mission/app/order/status/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.project.Week_Mission.app.order.controller;
+package com.project.Week_Mission.app.order.status;
 
 public enum OrderStatus {
